+++ dashboard/head.py
@@ -1,6 +1,5 @@
 import sys
 import socket
-import json
 import asyncio
 import logging
 import ipaddress
@@ -16,7 +15,8 @@ import ray.new_dashboard.utils as dashboard_utils
 from ray import ray_constants
 from ray.core.generated import gcs_service_pb2
 from ray.core.generated import gcs_service_pb2_grpc
-from ray.new_dashboard.datacenter import DataSource, DataOrganizer
+from ray.new_dashboard.datacenter import DataOrganizer
+from ray.new_dashboard.utils import async_loop_forever
 
 logger = logging.getLogger(__name__)
 routes = dashboard_utils.ClassMethodRouteTable
@@ -24,16 +24,12 @@ routes = dashboard_utils.ClassMethodRouteTable
 aiogrpc.init_grpc_aio()
 
 
-def gcs_node_info_to_dict(message):
-    return dashboard_utils.message_to_dict(
-        message, {"nodeId"}, including_default_value_fields=True)
-
-
 class DashboardHead:
     def __init__(self, http_host, http_port, http_port_retries, redis_address,
                  redis_password, log_dir):
-        # NodeInfoGcsService
-        self._gcs_node_info_stub = None
+        # HeartbeatInfoGcsService
+        self._gcs_heartbeat_info_stub = None
+        self._gcs_check_alive_seq = 0
         self._gcs_rpc_error_counter = 0
         # Public attributes are accessible for all head modules.
         # Walkaround for issue: https://github.com/ray-project/ray/issues/7084
@@ -52,75 +48,33 @@ class DashboardHead:
         logger.info("Dashboard head grpc address: %s:%s", self.ip,
                     self.grpc_port)
 
-    async def _get_nodes(self):
-        """Read the client table.
-
-        Returns:
-            A dict of information about the nodes in the cluster.
-        """
-        request = gcs_service_pb2.GetAllNodeInfoRequest()
-        reply = await self._gcs_node_info_stub.GetAllNodeInfo(
-            request, timeout=2)
-        if reply.status.code == 0:
-            result = {}
-            for node_info in reply.node_info_list:
-                node_info_dict = gcs_node_info_to_dict(node_info)
-                result[node_info_dict["nodeId"]] = node_info_dict
-            return result
-        else:
-            logger.error("Failed to GetAllNodeInfo: %s", reply.status.message)
-
-    async def _update_nodes(self):
-        while True:
-            try:
-                nodes = await self._get_nodes()
-
-                alive_node_ids = []
-                alive_node_infos = []
-                node_id_to_ip = {}
-                node_id_to_hostname = {}
-                for node in nodes.values():
-                    node_id = node["nodeId"]
-                    ip = node["nodeManagerAddress"]
-                    hostname = node["nodeManagerHostname"]
-                    node_id_to_ip[node_id] = ip
-                    node_id_to_hostname[node_id] = hostname
-                    assert node["state"] in ["ALIVE", "DEAD"]
-                    if node["state"] == "ALIVE":
-                        alive_node_ids.append(node_id)
-                        alive_node_infos.append(node)
-
-                agents = dict(DataSource.agents)
-                for node_id in alive_node_ids:
-                    key = f"{dashboard_consts.DASHBOARD_AGENT_PORT_PREFIX}" \
-                          f"{node_id}"
-                    agent_port = await self.aioredis_client.get(key)
-                    if agent_port:
-                        agents[node_id] = json.loads(agent_port)
-                for node_id in agents.keys() - set(alive_node_ids):
-                    agents.pop(node_id, None)
-
-                DataSource.node_id_to_ip.reset(node_id_to_ip)
-                DataSource.node_id_to_hostname.reset(node_id_to_hostname)
-                DataSource.agents.reset(agents)
-                DataSource.nodes.reset(nodes)
-
-                self._gcs_rpc_error_counter = 0
-            except aiogrpc.AioRpcError:
-                logger.exception("Got AioRpcError when updating nodes.")
-                self._gcs_rpc_error_counter += 1
-                if self._gcs_rpc_error_counter > \
-                        dashboard_consts.MAX_COUNT_OF_GCS_RPC_ERROR:
-                    logger.error(
-                        "Dashboard suicide, the GCS RPC error count %s > %s",
-                        self._gcs_rpc_error_counter,
-                        dashboard_consts.MAX_COUNT_OF_GCS_RPC_ERROR)
-                    sys.exit(-1)
-            except Exception:
-                logger.exception("Error updating nodes.")
-            finally:
-                await asyncio.sleep(
-                    dashboard_consts.UPDATE_NODES_INTERVAL_SECONDS)
+    @async_loop_forever(dashboard_consts.GCS_CHECK_ALIVE_INTERVAL_SECONDS)
+    async def _gcs_check_alive(self):
+        try:
+            self._gcs_check_alive_seq += 1
+            request = gcs_service_pb2.CheckAliveRequest(
+                seq=self._gcs_check_alive_seq)
+            reply = await self._gcs_heartbeat_info_stub.CheckAlive(
+                request, timeout=2)
+            if reply.status.code != 0:
+                raise Exception(
+                    f"Failed to CheckAlive: {reply.status.message}")
+            self._gcs_rpc_error_counter = 0
+        except aiogrpc.AioRpcError:
+            logger.exception(
+                "Got AioRpcError when checking GCS is alive, seq=%s.",
+                self._gcs_check_alive_seq)
+            self._gcs_rpc_error_counter += 1
+            if self._gcs_rpc_error_counter > \
+                    dashboard_consts.GCS_CHECK_ALIVE_MAX_COUNT_OF_RPC_ERROR:
+                logger.error(
+                    "Dashboard suicide, the GCS RPC error count %s > %s",
+                    self._gcs_rpc_error_counter,
+                    dashboard_consts.GCS_CHECK_ALIVE_MAX_COUNT_OF_RPC_ERROR)
+                sys.exit(-1)
+        except Exception:
+            logger.exception("Error checking GCS is alive, seq=%s.",
+                             self._gcs_check_alive_seq)
 
     def _load_modules(self):
         """Load dashboard head modules."""
@@ -172,9 +126,10 @@ class DashboardHead:
                 self.aiogrpc_gcs_channel = channel
                 break
 
-        # Create a NodeInfoGcsServiceStub.
-        self._gcs_node_info_stub = gcs_service_pb2_grpc.NodeInfoGcsServiceStub(
-            self.aiogrpc_gcs_channel)
+        # Create a HeartbeatInfoGcsServiceStub.
+        self._gcs_heartbeat_info_stub = \
+            gcs_service_pb2_grpc.HeartbeatInfoGcsServiceStub(
+                self.aiogrpc_gcs_channel)
 
         # Start a grpc asyncio server.
         await self.server.start()
@@ -233,7 +188,7 @@ class DashboardHead:
         # Freeze signal after all modules loaded.
         dashboard_utils.SignalManager.freeze()
         concurrent_tasks = [
-            self._update_nodes(),
+            self._gcs_check_alive(),
             _async_notify(),
             DataOrganizer.purge(),
             DataOrganizer.organize(),
