+++ dashboard/modules/node/node_consts.py
@@ -6,3 +6,4 @@ ACTOR_CHANNEL = "ACTOR"
 ERROR_INFO_UPDATE_INTERVAL_SECONDS = 5
 LOG_INFO_UPDATE_INTERVAL_SECONDS = 5
 NIL_NODE_ID = ray.NodeID.nil().hex()
+UPDATE_NODES_INTERVAL_SECONDS = 5
