+++ dashboard/modules/node/tests/test_node.py
@@ -8,8 +8,7 @@ import random
 import pytest
 import ray
 import threading
-import ray.new_dashboard.modules.stats_collector.stats_collector_consts \
-    as stats_collector_consts
+from ray.new_dashboard.modules.node import node_consts
 from datetime import datetime, timedelta
 from ray.cluster_utils import Cluster
 from ray.new_dashboard.tests.conftest import *  # noqa
@@ -18,6 +17,52 @@ from ray.test_utils import (format_web_url, wait_until_server_available,
                             wait_until_succeeded_without_exception)
 
 logger = logging.getLogger(__name__)
+
+
+def test_nodes_update(enable_test_module, ray_start_with_dashboard):
+    assert (wait_until_server_available(ray_start_with_dashboard["webui_url"])
+            is True)
+    webui_url = ray_start_with_dashboard["webui_url"]
+    webui_url = format_web_url(webui_url)
+
+    timeout_seconds = 10
+    start_time = time.time()
+    while True:
+        time.sleep(1)
+        try:
+            response = requests.get(webui_url + "/test/dump")
+            response.raise_for_status()
+            try:
+                dump_info = response.json()
+            except Exception as ex:
+                logger.info("failed response: %s", response.text)
+                raise ex
+            assert dump_info["result"] is True
+            dump_data = dump_info["data"]
+            assert len(dump_data["nodes"]) == 1
+            assert len(dump_data["agents"]) == 1
+            assert len(dump_data["nodeIdToIp"]) == 1
+            assert len(dump_data["nodeIdToHostname"]) == 1
+            assert dump_data["nodes"].keys() == dump_data[
+                "nodeIdToHostname"].keys()
+
+            response = requests.get(webui_url + "/test/notified_agents")
+            response.raise_for_status()
+            try:
+                notified_agents = response.json()
+            except Exception as ex:
+                logger.info("failed response: %s", response.text)
+                raise ex
+            assert notified_agents["result"] is True
+            notified_agents = notified_agents["data"]
+            assert len(notified_agents) == 1
+            assert notified_agents == dump_data["agents"]
+            break
+        except (AssertionError, requests.exceptions.ConnectionError) as e:
+            logger.info("Retry because of %s", e)
+        finally:
+            if time.time() > start_time + timeout_seconds:
+                raise Exception("Timed out while testing.")
 
 
 def test_node_info(disable_aiohttp_cache, ray_start_with_dashboard):
@@ -407,8 +452,7 @@ def test_nil_node(enable_test_module, disable_aiohttp_cache,
             response = requests.get(webui_url + "/test/dump?key=node_actors")
             response.raise_for_status()
             result = response.json()
-            assert stats_collector_consts.NIL_NODE_ID not in result["data"][
-                "nodeActors"]
+            assert node_consts.NIL_NODE_ID not in result["data"]["nodeActors"]
             break
         except Exception as ex:
             last_ex = ex
